+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to
 
 ## [Unreleased]
 
+- Add `.git` `SkipRule` to all targets to skip `.git` directory traversal.
 - Port glob matcher from `picomatch` to `zeptomatch`.
 - Move `Mopsgamer/view-ignored` to `view-ignored/view-ignored` + all related repos.
 - Remove `packageJsonExtractor` from `VSCE` target so `package.json`'s `files` array is not wrongly used to exclude extension files.